+++ .vscode/launch.json
@@ -25,8 +25,8 @@
             "console": "integratedTerminal",
             //"console": "externalTerminal",
             "consoleTitle": "ptvsd.server",
-            //"program": "${file}",
-            "program": "${workspaceFolder}/tests/test_data/testpkgs/pkg1/__main__.py",
+            "program": "${file}",
+            "logToFile": true,
         },
         {
             "name": "Attach [debugServer]",
@@ -34,6 +34,7 @@
             "request": "attach",
             "host": "localhost",
             "port": 5678,
+            "logToFile": true,
         },
     ]
 }
